design/cinematograph_convert/requirements.txt
cffi==1.14.4
chevron==0.13.1
pyaml==20.4.0
pycparser==2.20
PyYAML==5.3.1
six==1.15.0
SoundFile==0.10.3.post1
